requirements.txt
cairosvg
configs
einops
fastapi
ipython
kivy
matplotlib
moviepy==1.0.3
networkx
numpy
pandas
Pillow
pydantic
PyYAML
qwen_vl_utils
runpod
scikit_learn
scipy
Shapely
transformers==4.51.3
uvicorn
accelerate
requests
bitsandbytes
torchvision
huggingface_hub